add WiFi connectivity with a HTTP test
The WiFi connection is set up in 'boot.py', using the 'wifi_connect'
function in './lib/wifi.py'. It requires two arguments: the SSID and
password for the WiFi. The values of the SSID and the password are set
in the environment variable file 'env.json'. And the environment
variables are loaded with the 'get_env' function in './lib/env.py'.

The HTTP test is used in 'main.py'. It checks so a HTTP GET request is
able to be made to a URL. It uses the 'http_get_request' function in
'./lib/http.py'.

An 'env.json.example' file is added as an example of how 'env.json'
should look. The .gitignore is changed to reflect the name of the
environment variable file.